Refactor/tidy.
