+++ src/poolCacher.ts
@@ -1,5 +1,5 @@
 import cloneDeep from 'lodash.clonedeep';
-import { PoolDataService, SubgraphPoolBase } from './types';
+import { PoolDataService, SubgraphPoolBase, SubgraphToken } from './types';
 
 export class PoolCacher {
     private pools: SubgraphPoolBase[] = [];
@@ -11,8 +11,28 @@ export class PoolCacher {
         return this._finishedFetching;
     }
 
-    public getPools(): SubgraphPoolBase[] {
-        return cloneDeep(this.pools);
+    public getPools(useBpts?: boolean): SubgraphPoolBase[] {
+        const pools = cloneDeep(this.pools);
+        // If we use join/exit paths add the pool token to its token list
+        if (useBpts) {
+            for (const pool of pools) {
+                if (
+                    pool.poolType === 'Weighted' ||
+                    pool.poolType === 'Investment'
+                ) {
+                    const BptAsToken: SubgraphToken = {
+                        address: pool.address,
+                        balance: pool.totalShares,
+                        decimals: 18,
+                        priceRate: '1',
+                        weight: '0',
+                    };
+                    pool.tokens.push(BptAsToken);
+                    pool.tokensList.push(pool.address);
+                }
+            }
+        }
+        return pools;
     }
 
     /*
@@ -22,7 +42,6 @@ export class PoolCacher {
         try {
             this.pools = await this.poolDataService.getPools();
             this._finishedFetching = true;
-
             return true;
         } catch (err) {
             // On error clear all caches and return false so user knows to try again.
