+++ test/testData/weightedPools/joinExitPools.json
